move other package interface Entity
