+++ homework/kaydunov_denis/shopping_list_application_part5/src/main/java/github/kaydunovdenis/bean/Entity.java
@@ -1,4 +1,4 @@
-package github.kaydunovdenis.bean.product;
+package github.kaydunovdenis.bean;
 
 public interface Entity {
     String getName();
