+++ homework/kaydunov_denis/shopping_list_application_part5/src/main/java/github/kaydunovdenis/bean/product/Product.java
@@ -1,5 +1,6 @@
 package github.kaydunovdenis.bean.product;
 
+import github.kaydunovdenis.bean.Entity;
 import github.kaydunovdenis.service.ProductService;
 import lombok.Data;
