+++ homework/kaydunov_denis/shopping_list_application_part5/src/main/java/github/kaydunovdenis/repository/Repository.java
@@ -1,6 +1,6 @@
 package github.kaydunovdenis.repository;
 
-import github.kaydunovdenis.bean.product.Entity;
+import github.kaydunovdenis.bean.Entity;
 
 import java.util.Map;
